crawler/newsapi/src/requirements.txt
pymongo
python-dotenv
requests
apscheduler